requirements.txt
appdirs==1.4.0
gitdb2==2.0.0
GitPython==2.1.1
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
smmap2==2.0.1
